п

+++ src/main/java/ru/quard0r/base/repository/CategoryRepository.java
@@ -7,7 +7,7 @@ import ru.quard0r.base.entity.Category;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     Optional<Category> findByName(String name);
